+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The ability to turn on/off services via environment variable.
 - Default plans for Spanner, BigTable, and CloudSQL.
 - Whitelists for bindings so only certain "safe" roles can be chosen by end-users.
+- Automatic labeling of resources with organizaiton GUID, space GUID, and instance ID to BigQuery, CloudSQL, Spanner, and GCS.
 
 ### Deprecated
 - Running the service by executing the main executable. Use the `serve` sub-command instead.